+++ lib/td_dd_web/views/data_structure_version_view.ex
@@ -8,6 +8,8 @@ defmodule TdDdWeb.DataStructureVersionView do
   alias TdDdWeb.TagView
   alias TdDqWeb.ImplementationStructureView
 
+  @protected DataStructures.protected()
+
   def render("show.json", %{actions: actions} = assigns) do
     "show.json"
     |> render(Map.delete(assigns, :actions))
@@ -322,11 +324,32 @@ defmodule TdDdWeb.DataStructureVersionView do
   defp lift_data(attrs), do: attrs
 
   defp merge_metadata(%{metadata_versions: [_ | _] = metadata_versions} = dsv) do
-    %{fields: mutable_metadata} = Enum.max_by(metadata_versions, & &1.version)
+    %{fields: mutable_metadata} =
+      Enum.max_by(metadata_versions, & &1.version)
+      |> case do
+        %{deleted_at: deleted_at} = _version when deleted_at != nil ->
+          %{fields: %{}}
+
+        version ->
+          version
+      end
 
     Map.update(dsv, :metadata, mutable_metadata, fn
-      nil -> mutable_metadata
-      %{} = metadata -> Map.merge(metadata, mutable_metadata)
+      nil ->
+        mutable_metadata
+
+      %{} = metadata ->
+        Map.merge(
+          metadata,
+          mutable_metadata,
+          fn
+            # Merge metadata and mutable_metadata @protected fields.
+            # If there is a conflict in the @protected content, the one from
+            # mutable metadata (rightmost in the Map.merge) will prevail.
+            @protected, mp, mmp -> Map.merge(mp, mmp)
+            _key, _mp, mmp -> mmp
+          end
+        )
     end)
   end