+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - [TD-5076] Allow to clone implementations for non admin users
 - [TD-5169] Improve list_data_structure_versions since query performance
 - [TD-5170] Capture graph tests Logger.warn output
+- [TD-5082] Filter protected metadata (field "_protected")
+
 
 ### Changed